# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PubliMentor API
  version: 1.0.0
  description: |
    Scientific Editorial Workflow Management API.
    
    ## Overview
    PubliMentor provides APIs for managing manuscripts, finding reviewers,
    checking conflicts of interest, and validating research integrity.
    
    ## Authentication
    All endpoints except health checks require Bearer token authentication.
    
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    API requests are rate limited. When exceeded, responses include:
    - `429 Too Many Requests` status
    - `Retry-After` header with seconds until reset
    
    ## Request IDs
    All responses include `X-Request-ID` header for tracing.
    Include `X-Request-ID` in requests to correlate logs.
  contact:
    name: PubliMentor Support
    email: support@publimentor.io
  license:
    name: Proprietary
    url: https://publimentor.io/license
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.publimentor.io/api
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Create a new user account
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # Manuscript Endpoints
  # ============================================================
...
  /integrity/references:
    post:
      tags: [Integrity]
      summary: Validate references
      description: Validate reference DOIs and check for retractions
      operationId: validateReferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferenceValidationRequest'
      responses:
        '200':
          description: Reference validation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceValidationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /integrity/full-report:
    post:
      tags: [Integrity]
      summary: Generate comprehensive integrity report
      description: |
        Runs all integrity checks (tortured phrases, author identity, reference validation)
        in parallel and returns a unified report with overall risk assessment.
        Rate limited to 5 requests per minute per user.
      operationId: generateFullIntegrityReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  minLength: 100
                  description: Manuscript text for tortured phrase detection
                authors:
                  type: array
                  maxItems: 20
                  items:
                    type: object
                    required: [name]
                    properties:
                      name:
                        type: string
                      email:
                        type: string
                      orcid:
                        type: string
                      affiliation:
                        type: string
                referenceText:
                  type: string
                  minLength: 50
                  description: Raw reference list text for validation
                checks:
                  type: object
                  properties:
                    torturedPhrases:
                      type: boolean
                      default: true
                    authorIdentity:
                      type: boolean
                      default: true
                    references:
                      type: boolean
                      default: true
      responses:
        '200':
          description: Full integrity report
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportId:
                    type: string
                  generatedAt:
                    type: string
                    format: date-time
                  overallAssessment:
                    type: object
                    properties:
                      riskLevel:
                        type: string
                        enum: [clear, review, attention]
                      totalIndicators:
                        type: integer
                      checksCompleted:
                        type: integer
                      checksFailed:
                        type: integer
                  disclaimer:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

# ============================================================
# Components
# ============================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ApiResponse:
      type: object
      required: [success, requestId, timestamp]
      properties:
        success:
          type: boolean
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        data:
          type: object
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorResponse:
      type: object
      required: [success, requestId, timestamp, error]
      properties:
        success:
          type: boolean
          enum: [false]
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            target:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  target:
                    type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean

    # Health schemas
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
        name:
          type: string
        institution:
          type: string

    # Manuscript schemas
...
    ReferenceValidationRequest:
      type: object
      required: [references]
      properties:
        references:
          type: array
          items:
            type: object
            properties:
              doi:
                type: string
              pmid:
                type: string
              title:
                type: string
    ReferenceValidationResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                validCount:
                  type: integer
                invalidCount:
                  type: integer
                results:
                  type: array
                  items:
                    type: object
...
security:
  - bearerAuth: []